Moved benchmark and utils files to folders
